feat: Detect changes in property fields
The goal of this feature branch is to enhance the package's
change detection mechanism. Right now the only change detection
taking place is the default Angular one, i.e. when the input
property object reference changes. These changes improve on that
by using the `ngDoCheck` lifecycle hook to examine changes in
fields of input properties.